+++ lib/src/form_builder/email_textfield_parser.dart
@@ -94,7 +94,9 @@ class _EmailTextFieldCreatorState extends State<EmailTextFieldCreator> {
                 ? widget.map.conditional.show
                 : true
             : true;
-        if (widget.map.calculateValue != null && snapshot.data != null) {
+        if (widget.map.calculateValue != null &&
+            widget.map.calculateValue.isNotEmpty &&
+            snapshot.data != null) {
           _calculate = "";
           _keys.asMap().forEach((value, element) {
             widget.controller.text = (snapshot.data.containsKey(element))