+++ lib/src/form_builder/number_textfield_parser.dart
@@ -95,7 +95,9 @@ class _NumberTextFieldCreatorState extends State<NumberTextFieldCreator> {
                 ? widget.map.conditional.show
                 : true
             : true;
-        if (widget.map.calculateValue != null && snapshot.data != null) {
+        if (widget.map.calculateValue != null &&
+            widget.map.calculateValue.isNotEmpty &&
+            snapshot.data != null) {
           _calculate = "";
           _keys.asMap().forEach((value, element) {
             _calculate = (snapshot.data.containsKey(element))