+++ lib/src/form_builder/textarea_parser.dart
@@ -96,7 +96,9 @@ class _TextAreaCreatorState extends State<TextAreaCreator> {
                 : true
             : true;
 
-        if (widget.map.calculateValue != null && snapshot.data != null) {
+        if (widget.map.calculateValue != null &&
+            widget.map.calculateValue.isNotEmpty &&
+            snapshot.data != null) {
           _calculate = "";
           _keys.asMap().forEach(
             (value, element) {